+++ processor/src/test/resources/expected/com/example/observable/Arez_AbstractPrimitiveObservablesWithInitializerModel.java
@@ -1,0 +1,167 @@
+package com.example.observable;
+
+import arez.Arez;
+import arez.ArezContext;
+import arez.Component;
+import arez.Disposable;
+import arez.Observable;
+import arez.component.ComponentObservable;
+import arez.component.ComponentState;
+import arez.component.Identifiable;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import org.realityforge.braincheck.Guards;
+
+@Generated("arez.processor.ArezProcessor")
+public final class Arez_AbstractPrimitiveObservablesWithInitializerModel extends AbstractPrimitiveObservablesWithInitializerModel implements Disposable, Identifiable<Long>, ComponentObservable {
+  private static volatile long $$arezi$$_nextId;
+
+  private final long $$arezi$$_id;
+
+  private byte $$arezi$$_state;
+
+  @Nullable
+  private final ArezContext $$arezi$$_context;
+
+  private final Component $$arezi$$_component;
+
+  private final Observable<Boolean> $$arezi$$_disposedObservable;
+
+  @Nonnull
+  private final Observable<Long> $$arez$$_time;
+
+  private long $$arezd$$_time;
+
+  public Arez_AbstractPrimitiveObservablesWithInitializerModel(final long time) {
+    super();
+    this.$$arezd$$_time = time;
+    this.$$arezi$$_context = Arez.areZonesEnabled() ? Arez.context() : null;
+    this.$$arezi$$_id = ( Arez.areNativeComponentsEnabled() || Arez.areNativeComponentsEnabled() ) ? $$arezi$$_nextId++ : 0L;
+    this.$$arezi$$_state = ComponentState.COMPONENT_INITIALIZED;
+    this.$$arezi$$_component = Arez.areNativeComponentsEnabled() ? $$arezi$$_context().createComponent( "AbstractPrimitiveObservablesWithInitializerModel", $$arezi$$_id(), $$arezi$$_name() ) : null;
+    this.$$arezi$$_disposedObservable = $$arezi$$_context().createObservable( Arez.areNativeComponentsEnabled() ? this.$$arezi$$_component : null, Arez.areNamesEnabled() ? $$arezi$$_name() + ".isDisposed" : null, Arez.arePropertyIntrospectorsEnabled() ? () -> this.$$arezi$$_state >= 0 : null );
+    this.$$arez$$_time = $$arezi$$_context().createObservable( Arez.areNativeComponentsEnabled() ? this.$$arezi$$_component : null, Arez.areNamesEnabled() ? $$arezi$$_name() + ".time" : null, Arez.arePropertyIntrospectorsEnabled() ? () -> this.$$arezd$$_time : null, Arez.arePropertyIntrospectorsEnabled() ? v -> this.$$arezd$$_time = v : null );
+    this.$$arezi$$_state = ComponentState.COMPONENT_CONSTRUCTED;
+    if ( Arez.areNativeComponentsEnabled() ) {
+      this.$$arezi$$_component.complete();
+    }
+    this.$$arezi$$_state = ComponentState.COMPONENT_READY;
+  }
+
+  final ArezContext $$arezi$$_context() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> ComponentState.hasBeenInitialized( this.$$arezi$$_state ), () -> "Method named '$$arezi$$_context' invoked on uninitialized component of type 'AbstractPrimitiveObservablesWithInitializerModel'" );
+    }
+    return Arez.areZonesEnabled() ? this.$$arezi$$_context : Arez.context();
+  }
+
+  final long $$arezi$$_id() {
+    if ( Arez.shouldCheckInvariants() && !Arez.areNamesEnabled() && !Arez.areNativeComponentsEnabled() ) {
+      Guards.fail( () -> "Method invoked to access id when id not expected." );
+    }
+    return this.$$arezi$$_id;
+  }
+
+  @Override
+  @Nonnull
+  public final Long getArezId() {
+    return $$arezi$$_id();
+  }
+
+  String $$arezi$$_name() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> ComponentState.hasBeenInitialized( this.$$arezi$$_state ), () -> "Method named '$$arezi$$_name' invoked on uninitialized component of type 'AbstractPrimitiveObservablesWithInitializerModel'" );
+    }
+    return "AbstractPrimitiveObservablesWithInitializerModel." + $$arezi$$_id();
+  }
+
+  private boolean $$arezi$$_observe() {
+    final boolean isDisposed = isDisposed();
+    if ( !isDisposed )  {
+      this.$$arezi$$_disposedObservable.reportObserved();
+    }
+    return !isDisposed;
+  }
+
+  @Override
+  public boolean observe() {
+    return $$arezi$$_observe();
+  }
+
+  @Override
+  public boolean isDisposed() {
+    return ComponentState.isDisposingOrDisposed( this.$$arezi$$_state );
+  }
+
+  @Override
+  public void dispose() {
+    if ( !ComponentState.isDisposingOrDisposed( this.$$arezi$$_state ) ) {
+      this.$$arezi$$_state = ComponentState.COMPONENT_DISPOSING;
+      if ( Arez.areNativeComponentsEnabled() ) {
+        this.$$arezi$$_component.dispose();
+      } else {
+        $$arezi$$_context().dispose( Arez.areNamesEnabled() ? $$arezi$$_name() : null, () -> { {
+          this.$$arezi$$_disposedObservable.dispose();
+          this.$$arez$$_time.dispose();
+        } } );
+      }
+      this.$$arezi$$_state = ComponentState.COMPONENT_DISPOSED;
+    }
+  }
+
+  @Override
+  public long getTime() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> ComponentState.isActive( this.$$arezi$$_state ), () -> "Method named 'getTime' invoked on " + ComponentState.describe( this.$$arezi$$_state ) + " component named '" + $$arezi$$_name() + "'" );
+    }
+    this.$$arez$$_time.reportObserved();
+    return this.$$arezd$$_time;
+  }
+
+  @Override
+  public void setTime(final long value) {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> ComponentState.isActive( this.$$arezi$$_state ), () -> "Method named 'setTime' invoked on " + ComponentState.describe( this.$$arezi$$_state ) + " component named '" + $$arezi$$_name() + "'" );
+    }
+    if ( value != this.$$arezd$$_time ) {
+      this.$$arez$$_time.preReportChanged();
+      this.$$arezd$$_time = value;
+      this.$$arez$$_time.reportChanged();
+    }
+  }
+
+  @Override
+  public final int hashCode() {
+    if ( Arez.areNativeComponentsEnabled() ) {
+      return Long.hashCode( $$arezi$$_id() );
+    } else {
+      return super.hashCode();
+    }
+  }
+
+  @Override
+  public final boolean equals(final Object o) {
+    if ( Arez.areNativeComponentsEnabled() ) {
+      if ( this == o ) {
+        return true;
+      } else if ( null == o || !(o instanceof Arez_AbstractPrimitiveObservablesWithInitializerModel) ) {
+        return false;
+      } else {
+        final Arez_AbstractPrimitiveObservablesWithInitializerModel that = (Arez_AbstractPrimitiveObservablesWithInitializerModel) o;;
+        return $$arezi$$_id() == that.$$arezi$$_id();
+      }
+    } else {
+      return super.equals( o );
+    }
+  }
+
+  @Override
+  public final String toString() {
+    if ( Arez.areNamesEnabled() ) {
+      return "ArezComponent[" + $$arezi$$_name() + "]";
+    } else {
+      return super.toString();
+    }
+  }
+}