+++ processor/src/test/resources/input/com/example/observable/AbstractNonPrimitiveNonnullObservablesModel.java
@@ -1,0 +1,16 @@
+package com.example.observable;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Observable;
+import java.util.Date;
+import javax.annotation.Nonnull;
+
+@ArezComponent
+public abstract class AbstractNonPrimitiveNonnullObservablesModel
+{
+  @Observable
+  @Nonnull
+  public abstract Date getTime();
+
+  public abstract void setTime( @Nonnull Date value );
+}